Rename js-emoji; add list of emoji
