+++ client-src/common/services/emoji-renderer.js
