+++ client-src/index.js
@@ -22,8 +22,6 @@ import Privacy from './meta/components/privacy';
 import Terms from './meta/components/terms';
 import store from './redux/store';
 
-import './common/services/js-emoji';
-
 const App = React.createClass({
   render() {
     return <div>Not Found Sorry</div>;
